Upgrade to protobuf v3.9.0
Catches up with the same version used by chromium.
Android's //external/protobuf stays behind to v3.0.0-beta-3.
Unfortunately the bazel build files are broken there.
Thankfully, everything seems still API compatible.

Test: builds and passes CI tests
Bug: 132880619

+++ src/tracing/core/tracing_service_impl.cc
@@ -2218,7 +2218,7 @@ void TracingServiceImpl::SnapshotClocks(std::vector<TracePacket>* packets,
           static_cast<uint64_t>(base::FromPosixTimespec(clock.ts).count()));
     }
     protos::ClockSnapshot::Clock* c = clock_snapshot->add_clocks();
-    c->set_clock_id(clock.type);
+    c->set_clock_id(static_cast<uint32_t>(clock.type));
     c->set_timestamp(
         static_cast<uint64_t>(base::FromPosixTimespec(clock.ts).count()));
   }
